音乐播放器

+++ MusicPlayer/src/servlet/DeleteServlet.java
@@ -1,5 +1,6 @@
 package servlet;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import dao.MusicDao;
 import entity.Music;
 
@@ -25,25 +26,32 @@ public class DeleteServlet extends HttpServlet {
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("utf-8");
         resp.setContentType("application/json;charset=utf-8");
-
-        String idStr = req.getParameter("id");
-        int id = Integer.parseInt(idStr);
+        String strId = req.getParameter("id");
+        int id = Integer.parseInt(strId);
         MusicDao musicDao = new MusicDao();
-        musicDao.findMusicByID(id);
-        Music music = new Music();
-        int delete =  musicDao.deleteMusicById(id);
+        //删除之前查看当前 id 是否存在
+        Music music =  musicDao.findMusicByID(id);
+        if (music == null){
+            return;
+        }
+        int delete = musicDao.deleteMusicById(id);
         Map<String,Object> return_map = new HashMap<>();
-
+        System.out.println(music.getUrl());
         if (delete == 1){
-            File file = new File("E:\\java_code\\dataCode\\Java_code\\MusicPlayer\\web\\music"+music.getUrl() + ".mp3");
+            //代表数据库删除了，但是服务器上的音乐还存在
+            File file = new File("E:\\java_code\\dataCode\\Java_code\\MusicPlayer\\web\\" + music.getUrl() + ".mp3");
             if (file.delete()){
                 return_map.put("msg",true);
                 System.out.println("服务器删除成功！");
-            }else {
 
+            }else {
                 return_map.put("msg",false);
                 System.out.println("服务器删除失败！");
             }
+        }else {
+            return_map.put("msg",false);
         }
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.writeValue(resp.getWriter(),return_map);
     }
 }
